Added option to include command line flags
The run command in this workflow has been assumed to be a single
executable name (.e.g. ls or myscript.sh).  However, if command
line options/arguments need to be specified, then they need to be
passed to main.sh as well. Currently, the default behavior is to
convert any spaces and change them to _.  So, the simple solution
is to use a filter than converts any _ to spaces in the runcmd.
This works.  However, not in all cases because some users will
have an _ in the command name or the option name, for example:
my_script.sh takes_this_arg
...will be converted to
my script.sh takes this arg
...which is not good! This demo is currently only a simple
template so I'm leaving it here for now as there may be a
broader string-passing strategy implemented in the future.